Add XRegExp library to support Timestamp regexp in Firefox
As of now, Firefox does not yet support named capture groups
in RegExp objects. This is used in timestamp.ts where the
Timestamp.full pattern is currently built into

```javascript
^\s*([<\[](?<beginDate>\d{4}-\d{2}-\d{2})\s+[a-zA-Z]+(?:\s+(?<beginTimeBegin>\d{2}:\d{2})(?:-(?<beginTimeEnd>\d{2}:\d{2}))?)?[>\]])(?:--[<\[](?<endDate>\d{4}-\d{2}-\d{2})\s+[a-zA-Z]+(?:\s+(?<endTimeBegin>\d{2}:\d{2})(?:-(?<endTimeEnd>\d{2}:\d{2}))?)?[>\]])?\s*$
```

the XRegExp library provides cross-browser named capture groups
support but returns capture groups as an array. This commit
applies XRegExp for use only for the Timestamp patterns, and
unpacks the match groups by explicit array indexing.

In the www playground element in Firefox, this commit:
1. enables loading of the playground
2. successfully parses a full org-mode timestamp string, e.g.

```org
** Meeting in Amsterdam
   <2004-08-23 Mon 01:23-16:45>--<2006-11-02 Thu 20:00-22:00>
```